Modificaton of nlmfiltPhasor
Updating the nlmfiltPhasor. For every pixel, nlmfiltPhasor now calculates the norm of
the difference between the window and the original plot, instead of the
mse. This boosts the calculation speed by at least 8 times.

nlmfiltPhasor_division.m uses the inverse of the similarity as input. This ensures
that no matter what intensity range the original intensity value is, the
calculation makes sanece.
nlmfiltPhasor.m uses the negetive exponential coefficient of the
similarity as input. However, there is an exp_h term that might needs to
be adjusted.
nlmfiltPhasor_PCA uses PCA for calculation. However, PCA only helps with
computational speed when a large mask size is used.

nlmfiltPhasor_org uses mse for slimilarity calculation. The calculation
is slow.